FEAT: matching now uses similar interface as sets
Matching now uses the same 'doForTargetInSequence' construct as
addition and removal from sets. The class for matching takes an
instance of an OffTargetFinder class which hold all offtargets found
as well as the targets to match and any index structure present.
This should allow for easy use with multiple seperate sequences even
if not all sequences fit in memory at once.